+++ components/domains/AddSongScreen/index.tsx
@@ -31,6 +31,7 @@ import {
 import { convertDurationToSecond, getDurationText } from "~/store/room/utils";
 import type { PlaylistItem } from "~/types";
 import AddSongGuideScreen from "../AddSongGuideScreen";
+import RequestAccordion from "./RequestAccordion";
 
 const defaultEndPoint = process.env
   .NEXT_PUBLIC_SERVER_DEFAULT_END_POINT as string;
@@ -56,28 +57,13 @@ function AddSongScreen({ onClickBackButton }: AddSongScreenProps) {
   const [isValid, setIsValid] = useState(false);
   const [isError, setIsError] = useState(false);
   const [isLoading, setIsLoading] = useState(false);
-  const queryClient = useQueryClient();
-  const [proposedPlaylist, setProposedPlaylist] = useRecoilState(
-    proposedPlaylistAtomState
-  );
   const roomId = useRecoilValue(roomIdAtomState);
   const playlistId = useRecoilValue(playlistIdAtomState);
   const isHost = useRecoilValue(isHostAtomState);
 
-  const { data } = useGetPlaylistPendingItems(playlistId, isHost);
   const { publish: publishAddPlaylist } = useAddPlaylistItemRequest(roomId);
   const { publish: publishSendPlaylistRequest } =
     useSendPlaylistItemRequest(roomId);
-  const { publish: publishAcceptPlaylistItemRequest } =
-    useAcceptPlaylistItemRequest(roomId);
-  const { publish: publishDecinePlaylistItemRequest } =
-    useDeclinePlaylistItemRequest(roomId);
-
-  useEffect(() => {
-    if (data !== undefined) {
-      setProposedPlaylist(data);
-    }
-  }, [data, setProposedPlaylist]);
 
   useEffect(() => {
     setIsValid(false);
@@ -136,31 +122,6 @@ function AddSongScreen({ onClickBackButton }: AddSongScreenProps) {
     }
   };
 
-  const handleAcceptPlaylist = async (item: PlaylistItem) => {
-    setIsLoading(true);
-    await publishAcceptPlaylistItemRequest({
-      playlistId: item.playlistId,
-      playlistItemId: item.playlistItemId,
-    });
-    await setProposedPlaylist(
-      proposedPlaylist.filter(
-        (listItem) => listItem.playlistItemId !== item.playlistItemId
-      )
-    );
-    await queryClient.invalidateQueries(queryKeys.pendingPlaylist(playlistId));
-    setIsLoading(false);
-  };
-
-  const handleDeclinePlaylist = async (item: PlaylistItem) => {
-    setIsLoading(true);
-    await publishDecinePlaylistItemRequest({
-      playlistId: item.playlistId,
-      playlistItemId: item.playlistItemId,
-    });
-    await queryClient.invalidateQueries(queryKeys.pendingPlaylist(playlistId));
-    setIsLoading(false);
-  };
-
   return (
     <Layout screenColor="rgba(0, 0, 0, 0.85)">
       <Spacer type="vertical" style={{ height: "100%", overflowY: "auto" }}>
@@ -213,43 +174,7 @@ function AddSongScreen({ onClickBackButton }: AddSongScreenProps) {
               />
             </S.YoutubeWrapper>
           )}
-
-          {proposedPlaylist.length ? (
-            <S.ProposedMusicListCard hidden={!isHost}>
-              <S.CardHeader>
-                <strong>{proposedPlaylist.length}건</strong>의 신청된 노래가
-                있어요
-              </S.CardHeader>
-              <S.CardContent>
-                {proposedPlaylist.map((item, idx) => (
-                  <S.CardItem key={`${idx}-${item.videoId}`}>
-                    <Spacer type="vertical" style={{ marginRight: 12 }}>
-                      <S.MusicTitle>{item.title}</S.MusicTitle>
-                      <S.MusicArtist>
-                        {getDurationText(item.duration || 0)}
-                      </S.MusicArtist>
-                    </Spacer>
-                    <Spacer gap={8} align="center">
-                      <S.Button
-                        color="#007aff"
-                        onClick={() => handleAcceptPlaylist(item)}
-                      >
-                        추가
-                      </S.Button>
-                      <S.Button
-                        color=" #F54031"
-                        onClick={() => handleDeclinePlaylist(item)}
-                      >
-                        거절
-                      </S.Button>
-                    </Spacer>
-                  </S.CardItem>
-                ))}
-              </S.CardContent>
-            </S.ProposedMusicListCard>
-          ) : (
-            <></>
-          )}
+          <RequestAccordion />
         </Spacer>
 
         <BottomButton
@@ -333,76 +258,6 @@ const S = {
       height: 100%;
       clip-path: inset(0% 0% 0% 0% round 8px);
     }
-  `,
-
-  ProposedMusicListCard: styled.div<{ hidden: boolean }>`
-    display: ${(p) => (p.hidden ? "none" : "")};
-    margin: 20px 0%;
-    background: rgba(26, 26, 26, 0.9);
-    border-radius: 20px;
-    padding: 20px;
-  `,
-
-  CardHeader: styled.div`
-    color: white;
-    border-bottom: 1px solid rgba(255, 255, 255, 0.1);
-    padding-bottom: 20px;
-
-    font-size: 14px;
-    line-height: 17px;
-    /* identical to box height */
-
-    letter-spacing: -0.452636px;
-
-    & > strong {
-      font-weight: 800;
-    }
-  `,
-
-  CardContent: styled.ul`
-    margin-top: 16px;
-    display: flex;
-    flex-direction: column;
-    gap: 16px;
-  `,
-  CardItem: styled.li`
-    display: flex;
-    justify-content: space-between;
-  `,
-
-  MusicTitle: styled.div`
-    color: white;
-    font-weight: 600;
-    font-size: 12px;
-    line-height: 155%;
-    line-clamp: 2;
-    text-overflow: ellipsis;
-    overflow: hidden;
-    display: -webkit-box;
-    -webkit-box-orient: vertical;
-    -webkit-line-clamp: 2;
-  `,
-  MusicArtist: styled.div`
-    font-weight: 300;
-    font-size: 12px;
-    line-height: 155%;
-    letter-spacing: -0.405328px;
-
-    color: #959595;
-  `,
-
-  Button: styled.button<{ color: string }>`
-    cursor: pointer;
-    border: none;
-    padding: 9px 16px;
-    background: ${(p) => p.color};
-    border-radius: 6px;
-    font-weight: 600;
-    font-size: 12px;
-    line-height: 14px;
-    letter-spacing: -0.306564px;
-    color: #fff;
-    white-space: nowrap;
   `,
 };
